src/app/requirements.txt
fastapi
uvicorn
llama-cpp-python
piper-tts
argostranslate
